Return list of ints from sum2

+++ src/PIL/ImageStat.py
@@ -108,14 +108,14 @@ class Stat:
         return v
 
     @cached_property
-    def sum2(self) -> list[float]:
+    def sum2(self) -> list[int]:
         """Squared sum of all pixels for each band in the image."""
 
         v = []
         for i in range(0, len(self.h), 256):
-            sum2 = 0.0
+            sum2 = 0
             for j in range(256):
-                sum2 += (j**2) * float(self.h[i + j])
+                sum2 += (j**2) * self.h[i + j]
             v.append(sum2)
         return v
